+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_URL = http://localhost:3001
