+++ src/App.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import './App.css';
-import {Header, Body, Footer} from './components'
+import Header from './components/Header';
+import Body from './components/Body';
+import Footer from './components/Footer';
 
 function App() {
   return (
