+++ src/components/Footer.tsx
@@ -1,0 +1,19 @@
+const Footer = () => {
+    return (
+      
+        <footer>
+          <div className='footer-div'>
+            <div className='logo-div'>
+              <p>THE MOTOR CORP</p>
+            </div>
+            <div className='social-div'>
+              <p>Contact Us</p>
+              <p>&copy; 2023 The Motor Corp</p>
+            </div>
+          </div>
+        </footer>
+      
+    )
+  }
+
+  export default Footer;
