+++ src/components/Header.tsx
@@ -1,0 +1,23 @@
+const Header = () => {
+    return (
+        <div>
+            <header>
+                <div className='header-div'>
+                    <div className='title-div'>
+                        <p>THE MOTOR CORP</p>
+                    </div>
+                    <div className='services-div p-style'>
+                        <p>MOTORCYCLES</p>
+                        <p>ACCESSORIES</p>
+                    </div>
+                    <div className='users-div p-style'>
+                        <p>BOOK NOW</p>
+                        <p>SIGNUP</p>
+                    </div>
+                </div>
+            </header>
+        </div>
+    )
+}
+
+export default Header;
